Move ChunksToMatrix() into test-only file
and remove Spans which are not needed in testing.
